refactor: header, main, slider 분리
